updates from feedback
